+++ e2e/poa_test.go
@@ -2,7 +2,6 @@ package e2e
 
 import (
 	"context"
-	"fmt"
 	"testing"
 
 	"github.com/strangelove-ventures/interchaintest/v8"
@@ -11,8 +10,6 @@ import (
 	"github.com/strangelove-ventures/poa"
 	"github.com/strangelove-ventures/poa/e2e/helpers"
 	"github.com/stretchr/testify/require"
-
-	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 )
 
 func TestPOABase(t *testing.T) {
@@ -49,103 +46,7 @@ func TestPOABase(t *testing.T) {
 	// === Test Cases ===
 	testStakingDisabled(t, ctx, chain, validators, acc0, acc1)
 	testPowerErrors(t, ctx, chain, validators, incorrectUser, acc0)
-	testPending(t, ctx, chain, acc0)
-	testGovernance(t, ctx, chain, acc0, validators)
-	testUpdatePOAParams(t, ctx, chain, acc0, incorrectUser)
-}
-
-func testUpdatePOAParams(t *testing.T, ctx context.Context, chain *cosmos.CosmosChain, acc0, incorrectUser ibc.Wallet) {
-	var tx helpers.TxResponse
-	var err error
-
-	t.Run("fail: update-params message from a non authorized user", func(t *testing.T) {
-		tx, err = helpers.POAUpdateParams(t, ctx, chain, incorrectUser, []string{incorrectUser.FormattedAddress()}, true)
-		if err != nil {
-			t.Fatal(err)
-		}
-		txRes, err := chain.GetTransaction(tx.Txhash)
-		require.NoError(t, err)
-		fmt.Printf("%+v", txRes)
-		require.Contains(t, txRes.RawLog, poa.ErrNotAnAuthority.Error())
-	})
-
-	t.Run("fail: update staking params from a non authorized user", func(t *testing.T) {
-		tx, err = helpers.POAUpdateStakingParams(t, ctx, chain, incorrectUser, stakingtypes.DefaultParams())
-		if err != nil {
-			t.Fatal(err)
-		}
-		txRes, err := chain.GetTransaction(tx.Txhash)
-		require.NoError(t, err)
-		fmt.Printf("%+v", txRes)
-		require.EqualValues(t, txRes.Code, 3)
-
-		sp := helpers.GetStakingParams(t, ctx, chain)
-		fmt.Printf("%+v", sp)
-	})
-
-	t.Run("success: update staking params from an authorized user.", func(t *testing.T) {
-		stakingparams := stakingtypes.DefaultParams()
-		tx, err = helpers.POAUpdateStakingParams(t, ctx, chain, acc0, stakingtypes.Params{
-			UnbondingTime:     stakingparams.UnbondingTime,
-			MaxValidators:     10,
-			MaxEntries:        stakingparams.MaxEntries,
-			HistoricalEntries: stakingparams.HistoricalEntries,
-			BondDenom:         stakingparams.BondDenom,
-			MinCommissionRate: stakingparams.MinCommissionRate,
-		})
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		txRes, err := chain.GetTransaction(tx.Txhash)
-		require.NoError(t, err)
-		fmt.Printf("%+v", txRes)
-		require.EqualValues(t, txRes.Code, 0)
-
-		sp := helpers.GetStakingParams(t, ctx, chain)
-		fmt.Printf("%+v", sp)
-		require.EqualValues(t, sp.MaxValidators, 10)
-	})
-
-	t.Run("success: update-params message from an authorized user.", func(t *testing.T) {
-		govModule := "cosmos10d07y265gmmuvt4z0w9aw880jnsr700j6zn9kn"
-		randAcc := "cosmos1qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqnrql8a"
-
-		newAdmins := []string{acc0.FormattedAddress(), govModule, randAcc, incorrectUser.FormattedAddress()}
-		tx, err = helpers.POAUpdateParams(t, ctx, chain, acc0, newAdmins, true)
-		if err != nil {
-			t.Fatal(err)
-		}
-		txRes, err := chain.GetTransaction(tx.Txhash)
-		require.NoError(t, err)
-		fmt.Printf("%+v", txRes)
-		require.EqualValues(t, txRes.Code, 0)
-
-		p := helpers.GetPOAParams(t, ctx, chain)
-		for _, admin := range newAdmins {
-			require.Contains(t, p.Admins, admin)
-		}
-	})
-
-	t.Run("success: gov proposal update", func(t *testing.T) {
-		govModule := "cosmos10d07y265gmmuvt4z0w9aw880jnsr700j6zn9kn"
-		randAcc := "cosmos1qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqnrql8a"
-
-		updatedParams := []cosmos.ProtoMessage{
-			&poa.MsgUpdateParams{
-				Sender: govModule,
-				Params: poa.Params{
-					Admins: []string{acc0.FormattedAddress(), govModule, randAcc},
-				},
-			},
-		}
-		propId := helpers.SubmitParamChangeProp(t, ctx, chain, incorrectUser, updatedParams, govModule, 25)
-		helpers.ValidatorVote(t, ctx, chain, propId, cosmos.ProposalVoteYes, 30)
-		for _, admin := range helpers.GetPOAParams(t, ctx, chain).Admins {
-			require.NotEqual(t, admin, incorrectUser.FormattedAddress())
-		}
-	})
-
+	testRemovePending(t, ctx, chain, acc0)
 }
 
 func testStakingDisabled(t *testing.T, ctx context.Context, chain *cosmos.CosmosChain, validators []string, acc0, acc1 ibc.Wallet) {
@@ -172,39 +73,20 @@ func testStakingDisabled(t *testing.T, ctx context.Context, chain *cosmos.Cosmos
 	require.Contains(t, res.RawLog, poa.ErrStakingActionNotAllowed.Error())
 }
 
-func testPending(t *testing.T, ctx context.Context, chain *cosmos.CosmosChain, acc0 ibc.Wallet) {
+func testRemovePending(t *testing.T, ctx context.Context, chain *cosmos.CosmosChain, acc0 ibc.Wallet) {
 	t.Log("\n===== TEST PENDING =====")
 
 	_, err := helpers.POACreatePendingValidator(t, ctx, chain, acc0, "pl3Q8OQwtC7G2dSqRqsUrO5VZul7l40I+MKUcejqRsg=", "testval", "0.10", "0.25", "0.05")
 	require.NoError(t, err)
 
-	pv := helpers.GetPOAPending(t, ctx, chain)
-	require.Equal(t, 1, len(pv.Pending))
-	require.Equal(t, "0", pv.Pending[0].Tokens)
-	require.Equal(t, "1", pv.Pending[0].MinSelfDelegation)
+	pv := helpers.GetPOAPending(t, ctx, chain).Pending
+	require.Equal(t, 1, len(pv))
 
-	_, err = helpers.POARemovePending(t, ctx, chain, acc0, pv.Pending[0].OperatorAddress)
+	_, err = helpers.POARemovePending(t, ctx, chain, acc0, pv[0].OperatorAddress)
 	require.NoError(t, err)
 
 	// validate it was removed
-	pv = helpers.GetPOAPending(t, ctx, chain)
-	require.Equal(t, 0, len(pv.Pending))
-}
-
-func testGovernance(t *testing.T, ctx context.Context, chain *cosmos.CosmosChain, acc0 ibc.Wallet, validators []string) {
-	t.Log("\n===== TEST GOVERNANCE =====")
-
-	authorityAddr := "cosmos10d07y265gmmuvt4z0w9aw880jnsr700j6zn9kn" // gov
-
-	// ibc.ChainConfig key: app_state.poa.params.admins must contain the governance address.
-	propID := helpers.SubmitGovernanceProposalForValidatorChanges(t, ctx, chain, acc0, validators[0], 1_234_567, true, authorityAddr)
-	helpers.ValidatorVote(t, ctx, chain, propID, cosmos.ProposalVoteYes, 25)
-
-	// validate the validator[0] was set to 1_234_567
-	val := helpers.GetValidators(t, ctx, chain).Validators[0]
-	require.Equal(t, val.Tokens, "1234567")
-	p := helpers.GetPOAConsensusPower(t, ctx, chain, val.OperatorAddress)
-	require.EqualValues(t, 1_234_567/1_000_000, p)
+	require.Equal(t, 0, len(helpers.GetPOAPending(t, ctx, chain).Pending))
 }
 
 func testPowerErrors(t *testing.T, ctx context.Context, chain *cosmos.CosmosChain, validators []string, incorrectUser ibc.Wallet, admin ibc.Wallet) {
@@ -232,19 +114,4 @@ func testPowerErrors(t *testing.T, ctx context.Context, chain *cosmos.CosmosChai
 		require.NoError(t, err)
 		require.Contains(t, res.RawLog, poa.ErrUnsafePower.Error())
 	})
-}
-
-// assertSignatures asserts that the current block has the exact number of signatures as expected
-func assertSignatures(t *testing.T, ctx context.Context, chain *cosmos.CosmosChain, expectedSigs int) {
-	height, err := chain.GetNode().Height(ctx)
-	require.NoError(t, err)
-	block := helpers.GetBlockData(t, ctx, chain, height)
-	require.Equal(t, expectedSigs, len(block.LastCommit.Signatures))
-
-}
-
-func assertConsensus(t *testing.T, ctx context.Context, chain *cosmos.CosmosChain, expected int) {
-	cbft := helpers.GetCometBFTConsensus(t, ctx, chain)
-	amt := len(cbft.Validators)
-	require.EqualValues(t, amt, expected, "expected %d in consensus, got %d", expected, amt)
 }
